Move database tests to instrumented
