+++ collect_app/src/androidTest/java/org/odk/collect/android/instrumented/database/helpers/InstancesDatabaseHelperTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.database.helpers;
+package org.odk.collect.android.instrumented.database.helpers;
 
 import android.database.sqlite.SQLiteDatabase;
 
@@ -9,6 +9,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.odk.collect.android.dao.InstancesDao;
+import org.odk.collect.android.database.helpers.InstancesDatabaseHelper;
 import org.odk.collect.android.instances.Instance;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.android.utilities.SQLiteUtils;
