+++ collect_app/src/main/java/org/odk/collect/android/database/helpers/InstancesDatabaseHelper.java
@@ -53,7 +53,7 @@ public class InstancesDatabaseHelper extends SQLiteOpenHelper {
     private static final String DATABASE_NAME = "instances.db";
     public static final String INSTANCES_TABLE_NAME = "instances";
 
-    static final int DATABASE_VERSION = 6;
+    public static final int DATABASE_VERSION = 6;
 
     private static final String[] COLUMN_NAMES_V5 = {_ID, DISPLAY_NAME, SUBMISSION_URI, CAN_EDIT_WHEN_COMPLETE,
             INSTANCE_FILE_PATH, JR_FORM_ID, JR_VERSION, STATUS, LAST_STATUS_CHANGE_DATE, DELETED_DATE};
@@ -62,7 +62,7 @@ public class InstancesDatabaseHelper extends SQLiteOpenHelper {
         CAN_EDIT_WHEN_COMPLETE, INSTANCE_FILE_PATH, JR_FORM_ID, JR_VERSION, STATUS,
         LAST_STATUS_CHANGE_DATE, DELETED_DATE, GEOMETRY, GEOMETRY_TYPE};
 
-    static final String[] CURRENT_VERSION_COLUMN_NAMES = COLUMN_NAMES_V6;
+    public static final String[] CURRENT_VERSION_COLUMN_NAMES = COLUMN_NAMES_V6;
 
     private static boolean isDatabaseBeingMigrated;
